pyproject.toml
[project]
name = "ws-image-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]
